Added token endpoint, and more tests
Currently, i am trying to use an in mem-db to mock the db for tests.
So tests are unstable, but running as production profile works. This
is still very much under WIP, the register doesnt handle any proper
business logic, and neither does the token generator endpoint. I am
going to add proper business logic once i get the in-mem db to work.